replace string compare for enum deserialization with the same logic as for complex types
