+++ SpanJson/Helpers/MemberComparisonBuilder.cs
@@ -1,0 +1,220 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Text;
+using SpanJson.Resolvers;
+
+namespace SpanJson.Helpers
+{
+    public static class MemberComparisonBuilder
+    {
+        private static int GetSymbolSize<TSymbol>() where TSymbol : struct
+        {
+
+            if (typeof(TSymbol) == typeof(char))
+            {
+                return sizeof(char);
+            }
+
+            if (typeof(TSymbol) == typeof(byte))
+            {
+                return sizeof(byte);
+            }
+
+            throw new NotSupportedException();
+        }
+
+        public static Expression Build<TSymbol>(List<JsonMemberInfo> memberInfos, int index, ParameterExpression lengthParameter,
+            ParameterExpression nameSpanExpression, LabelTarget endOfBlockLabel,
+            Func<JsonMemberInfo, Expression> matchExpressionFunctor) where TSymbol : struct
+        {
+            var symbolSize = GetSymbolSize<TSymbol>();
+            var grouping = memberInfos.Where(a => a.CanRead && GetLength<TSymbol>(a.Name) >= index).GroupBy(a => CalculateKey<TSymbol>(a.Name, index))
+                .OrderByDescending(a => a.Key.Key).ToList();
+            if (!grouping.Any())
+            {
+                throw new InvalidOperationException(); // should never happen
+            }
+
+            var expressions = new List<Expression>();
+            foreach (var group in grouping)
+            {
+                if (group.Count() == 1) // need to check remaining values too 
+                {
+                    var memberInfo = group.Single();
+                    var length = index + group.Key.offset;
+                    var matchExpression = matchExpressionFunctor(memberInfo);
+                    matchExpression = Expression.Block(matchExpression, Expression.Goto(endOfBlockLabel));
+                    Expression comparisonExpression = null;
+                    if (group.Key.Key != 0 || group.Key.offset != 0)
+                    {
+                        comparisonExpression = Expression.Equal(GetReadMethod(nameSpanExpression, group.Key.intType, Expression.Constant(index * symbolSize)),
+                            GetConstantExpressionForGroupKey(group.Key.Key, group.Key.intType));
+                        var nameLength = GetLength<TSymbol>(memberInfo.Name);
+                        while (length < nameLength)
+                        {
+                            var subKey = CalculateKey<TSymbol>(memberInfo.Name, length);
+                            comparisonExpression = Expression.AndAlso(comparisonExpression,
+                                Expression.Equal(GetReadMethod(nameSpanExpression, subKey.intType, Expression.Constant(length * symbolSize)),
+                                    GetConstantExpressionForGroupKey(subKey.Key, subKey.intType)));
+                            length += subKey.offset;
+                        }
+                    }
+
+                    var lengthExpression = Expression.Equal(lengthParameter, Expression.Constant(length));
+                    var ifExpression = comparisonExpression == null ? lengthExpression : Expression.AndAlso(lengthExpression, comparisonExpression);
+                    expressions.Add(Expression.IfThen(ifExpression, matchExpression));
+                }
+                else
+                {
+                    var nextLength = index + group.Key.offset;
+                    var ifExpression = Expression.AndAlso(Expression.GreaterThanOrEqual(lengthParameter, Expression.Constant(nextLength)),
+                        Expression.Equal(GetReadMethod(nameSpanExpression, group.Key.intType, Expression.Constant(index * symbolSize)),
+                            GetConstantExpressionForGroupKey(group.Key.Key, group.Key.intType)));
+                    var subBlock = Build<TSymbol>(group.ToList(), nextLength, lengthParameter, nameSpanExpression,
+                        endOfBlockLabel, matchExpressionFunctor);
+                    expressions.Add(Expression.IfThen(ifExpression, subBlock));
+                }
+            }
+
+            return Expression.Block(expressions);
+        }
+
+        private static int GetLength<TSymbol>(string name)
+        {
+            if (typeof(TSymbol) == typeof(char))
+            {
+                return name.Length;
+            }
+
+            if (typeof(TSymbol) == typeof(byte))
+            {
+                return Encoding.UTF8.GetByteCount(name);
+            }
+
+            throw new NotSupportedException();
+        }
+
+        private static Expression GetConstantExpressionForGroupKey(ulong key, Type intType)
+        {
+            if (intType == typeof(byte))
+            {
+                return Expression.Constant((byte) key);
+            }
+
+            if (intType == typeof(ushort))
+            {
+                return Expression.Constant((ushort) key);
+            }
+
+            if (intType == typeof(uint))
+            {
+                return Expression.Constant((uint) key);
+            }
+
+            if (intType == typeof(ulong))
+            {
+                return Expression.Constant(key);
+            }
+
+            throw new NotSupportedException();
+        }
+
+        private static Expression GetReadMethod(ParameterExpression nameSpanExpression, Type intType, Expression offsetParameter)
+        {
+            string methodName;
+            if (intType == typeof(byte))
+            {
+                methodName = nameof(SpanHelper.ReadByte);
+            }
+
+            else if (intType == typeof(ushort))
+            {
+                methodName = nameof(SpanHelper.ReadUInt16);
+            }
+
+            else if (intType == typeof(uint))
+            {
+                methodName = nameof(SpanHelper.ReadUInt32);
+            }
+
+            else if (intType == typeof(ulong))
+            {
+                methodName = nameof(SpanHelper.ReadUInt64);
+            }
+            else
+            {
+                throw new NotSupportedException();
+            }
+
+            var methodInfo = typeof(SpanHelper).GetMethod(methodName, BindingFlags.Static | BindingFlags.Public);
+            return Expression.Call(methodInfo, nameSpanExpression, offsetParameter);
+        }
+
+        private static (ulong Key, Type intType, int offset) CalculateKey<TSymbol>(string memberName, int index)
+        {
+            if (typeof(TSymbol) == typeof(char))
+            {
+                return CalculateKeyUtf16(memberName, index); // for calculating the key the index is actually only half as much due to two byte chars
+            }
+
+            if (typeof(TSymbol) == typeof(byte))
+            {
+                return CalculateKeyUtf8(memberName, index);
+            }
+
+            throw new NotSupportedException();
+        }
+
+        private static (ulong Key, Type intType, int offset) CalculateKeyUtf8(string memberName, int index)
+        {
+            int remaining = GetLength<byte>(memberName) - index;
+            if (remaining >= 8)
+            {
+                return (BitConverter.ToUInt64(Encoding.UTF8.GetBytes(memberName), index), typeof(ulong), 8);
+            }
+
+            if (remaining >= 4)
+            {
+                return (BitConverter.ToUInt32(Encoding.UTF8.GetBytes(memberName), index), typeof(uint), 4);
+            }
+
+            if (remaining >= 2)
+            {
+                return (BitConverter.ToUInt16(Encoding.UTF8.GetBytes(memberName), index), typeof(ushort), 2);
+            }
+
+            if (remaining >= 1)
+            {
+                return (Encoding.UTF8.GetBytes(memberName)[index], typeof(byte), 1);
+            }
+
+            return (0, typeof(uint), 0);
+        }
+
+
+        private static (ulong Key, Type intType, int offset) CalculateKeyUtf16(string memberName, int index)
+        {
+            int remaining = GetLength<char>(memberName) - index;
+            if (remaining >= 4)
+            {
+                return (BitConverter.ToUInt64(Encoding.Unicode.GetBytes(memberName), index * 2), typeof(ulong), 4);
+            }
+
+            if (remaining >= 2)
+            {
+                return (BitConverter.ToUInt32(Encoding.Unicode.GetBytes(memberName), index * 2), typeof(uint), 2);
+            }
+
+            if (remaining >= 1)
+            {
+                return (BitConverter.ToUInt16(Encoding.Unicode.GetBytes(memberName), index * 2), typeof(ushort), 1);
+            }
+
+            return (0, typeof(uint), 0);
+        }
+    }
+}
